+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Perf
 
+### 2025-06-26
+
+- Refactor jump opcodes to use a blacklist on invalid targets.
+
 ### 2025-06-20
 
 - Use a lookup table for opcode parsing [#3253](https://github.com/lambdaclass/ethrex/pull/3253)